@@ -213,25 +215,26 @@
     "node-notifier": "^10.0.1",
     "open": "^10.1.0",
     "ora": "^8.1.0",
-    "pixelmatch": "^6.0.0",
+    "pixelmatch": "^7.1.0",
     "pngjs": "^7.0.0",
     "prettier": "^3.3.3",
     "prettier-plugin-jsdoc": "^1.3.0",
     "pretty-format": "^29.7.0",
     "prismjs": "^1.29.0",
-    "puppeteer": "^23.3.0",
+    "puppeteer": "^24.0.0",
     "qs": "^6.13.0",
-    "rc-drawer": "^7.2.0",
+    "rc-drawer": "^8.0.0",
     "rc-footer": "^0.6.8",
     "rc-resize-observer": "^1.4.0",
     "rc-virtual-list": "^3.14.5",
-    "react": "^18.3.1",
+    "react": "^19.0.0",
     "react-copy-to-clipboard": "^5.1.0",
     "react-countup": "^6.5.3",
-    "react-dom": "^18.3.1",
+    "react-dom": "^19.0.0",
     "react-highlight-words": "^0.20.0",
     "react-infinite-scroll-component": "^6.1.0",
     "react-intersection-observer": "^9.13.1",
+    "react-intl": "7.1.6",
     "react-resizable": "^3.0.5",
     "react-router-dom": "^7.0.1",
     "react-sticky-box": "^2.0.5",
