@@ -141,11 +141,11 @@ const PrevAndNext: React.FC<{ rtl?: boolean }> = ({ rtl }) => {
   return (
     <section className={styles.prevNextNav}>
       {prev &&
-        React.cloneElement(prev.label as ReactElement, {
+        React.cloneElement(prev.label as ReactElement<any>, {
           className: classNames(styles.pageNav, styles.prevNav, prev.className),
         })}
       {next &&
-        React.cloneElement(next.label as ReactElement, {
+        React.cloneElement(next.label as ReactElement<any>, {
           className: classNames(styles.pageNav, styles.nextNav, next.className),
         })}
     </section>
