@@ -35,7 +35,7 @@ const DocLayout: React.FC = () => {
   const location = useLocation();
   const { pathname, search, hash } = location;
   const [locale, lang] = useLocale(locales);
-  const timerRef = useRef<ReturnType<typeof setTimeout>>();
+  const timerRef = useRef<ReturnType<typeof setTimeout>>(null);
   const { direction } = useContext(SiteContext);
   const { loading } = useSiteData();
 
@@ -52,7 +52,11 @@ const DocLayout: React.FC = () => {
     timerRef.current = setTimeout(() => {
       nprogressHiddenStyle?.remove();
     }, 0);
-    return () => clearTimeout(timerRef.current);
+    return () => {
+      if (timerRef.current) {
+        clearTimeout(timerRef.current);
+      }
+    };
   }, []);
 
   // handle hash change or visit page hash from Link component, and jump after async chunk loaded
