@@ -27,9 +27,9 @@ export type { WelcomeProps } from './welcome';
 export { default as XProvider } from './x-provider';
 export type { XProviderProps } from './x-provider';
 
-export { default as useXChat } from './useXChat';
+export { default as useXChat } from './use-x-chat';
 
-export { default as useXAgent } from './useXAgent';
+export { default as useXAgent } from './use-x-agent';
 
 export { default as XStream } from './x-stream';
 export type { XStreamOptions } from './x-stream';
