@@ -35,7 +35,7 @@ const Demo: React.FC = () => {
         setLoading(false);
       }}
       actions={(_, info) => {
-        const { SendButton, LoadingButton, ClearButton } = info.components;
+        const { SendButton, LoadingButton, ClearButton, SpeechButton } = info.components;
 
         return (
           <Space size="small">
@@ -43,6 +43,7 @@ const Demo: React.FC = () => {
               <small>`Shift + Enter` to submit</small>
             </Typography.Text>
             <ClearButton />
+            <SpeechButton />
             {loading ? (
               <LoadingButton type="default" icon={<Spin size="small" />} disabled />
             ) : (
