@@ -17,12 +17,13 @@ import useStyle from './style';
 import useSpeech, { type AllowSpeech } from './useSpeech';
 
 import type { InputRef as AntdInputRef, ButtonProps, GetProps } from 'antd';
-import type { CustomizeComponent, SubmitType } from './interface';
+
+export type SubmitType = 'enter' | 'shiftEnter' | false;
 
 type TextareaProps = GetProps<typeof Input.TextArea>;
 
 export interface SenderComponents {
-  input?: CustomizeComponent<TextareaProps>;
+  input?: React.ComponentType<TextareaProps>;
 }
 
 export type ActionsRender = (
@@ -32,6 +33,7 @@ export type ActionsRender = (
       SendButton: React.ComponentType<ButtonProps>;
       ClearButton: React.ComponentType<ButtonProps>;
       LoadingButton: React.ComponentType<ButtonProps>;
+      SpeechButton: React.ComponentType<ButtonProps>;
     };
   },
 ) => React.ReactNode;
@@ -52,7 +54,7 @@ export interface SenderProps extends Pick<TextareaProps, 'placeholder' | 'onKeyP
   onCancel?: VoidFunction;
   onKeyDown?: React.KeyboardEventHandler<any>;
   onPaste?: React.ClipboardEventHandler<HTMLElement>;
-  onPasteFile?: (file: File) => void;
+  onPasteFile?: (firstFile: File, files: FileList) => void;
   components?: SenderComponents;
   styles?: {
     prefix?: React.CSSProperties;
@@ -80,8 +82,8 @@ export type SenderRef = {
 function getComponent<T>(
   components: SenderComponents | undefined,
   path: string[],
-  defaultComponent: CustomizeComponent<T>,
-): CustomizeComponent<T> {
+  defaultComponent: React.ComponentType<T>,
+): React.ComponentType<T> {
   return getValue(components, path) || defaultComponent;
 }
 
@@ -231,10 +233,10 @@ const ForwardSender = React.forwardRef<SenderRef, SenderProps>((props, ref) => {
 
   // ============================ Paste =============================
   const onInternalPaste: React.ClipboardEventHandler<HTMLElement> = (e) => {
-    // Get file
-    const file = e.clipboardData?.files[0];
-    if (file && onPasteFile) {
-      onPasteFile(file);
+    // Get files
+    const files = e.clipboardData?.files;
+    if (files?.length && onPasteFile) {
+      onPasteFile(files[0], files);
       e.preventDefault();
     }
 
@@ -269,6 +271,7 @@ const ForwardSender = React.forwardRef<SenderRef, SenderProps>((props, ref) => {
         SendButton,
         ClearButton,
         LoadingButton,
+        SpeechButton,
       },
     });
   } else if (actions) {
